src/api/requirements.txt
fastapi
fastapi[standard]
openai
requests
prompty[azure]==0.1.40
promptflow
azure-ai-evaluation
pandas
opentelemetry-instrumentation-fastapi
jsonlines
python-dotenv==1.0.1
azure-search-documents==11.4.0
gunicorn==21.2.0
jupyter
opentelemetry-sdk
opentelemetry-instrumentation
azure-identity==1.19.0
gunicorn==21.2.0
azure-keyvault-secrets
aiohttp==3.10.11
azure-cognitiveservices-speech==1.37.0
nbconvert
azure-ai-evaluation[remote]
validators
azureml-mlflow
azure-monitor-opentelemetry
opentelemetry-instrumentation-openai
azure-ai-inference[opentelemetry]
azure-ai-projects==1.0.0b4
azure-ai-contentsafety
pyodbc==5.2.0
